Refactored out app configuration from index file
